backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.13
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-extra-fields==3.7.0
filetype==1.2.0
gunicorn==20.1.0
idna==3.7
oauthlib==3.2.2
pillow==10.4.0
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.2
reportlab==4.2.5